Cast statfs file system type to unsigned long for switch
The file system type returned by statfs, on some 32-bit platforms,
is a signed value. But some of the file system magic types are,
in fact, large enough to be negative values. The usual hex value
used for, e.g. CIFS_MAGIC_NUMBER, does not cast to signed int on
newer compilers without disabling a warning. Casting the switch
value to unsigned long will allow these large magic values to
be compared without disabling any warnings.

+++ tools/tracegen/Port.cpp
@@ -684,7 +684,7 @@ Port::omrfile_stat(const char *path, unsigned int flags, struct J9FileStat *buf)
 	if (statfs(path, &statfsbuf)) {
 		return RC_FAILED;
 	}
-	switch (statfsbuf.f_type) {
+	switch ((unsigned long)(statfsbuf.f_type)) {
 	/* Detect remote filesystem types */
 	case 0x6969: /* NFS_SUPER_MAGIC */
 	case 0x517B: /* SMB_SUPER_MAGIC */
